Add logging to build.sh
Each time build.sh is run, a log is created in $WORKDIR/buildlogs/ with the directory name incorporating the current date-time.  Within this session-specific log directory is a single log file to which is appended all data dumped during the course of the script.

Convention will be that data echoed by the build script or other scripts written by Combine devs will be prepended with "### BUILDLOG:" and any other app's output would (obviously) not include this.

Output to the terminal in which the build.sh script is run will be limited due to the output redirection to log files.  This version of the script also tells the user at invocation the shell command needed to track verbose output of the install in real time.

Some output is coming from scripts run within containers, and is a bit trickier to deal with...specifically the combine/combine_db_prepare.sh script.